+++ src/main/java/kr/co/csalgo/presentation/subscription/SubscriptionController.java
@@ -1,4 +1,4 @@
-package kr.co.csalgo.presentation.user;
+package kr.co.csalgo.presentation.subscription;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
